# Optimization of the ToFieldIdentifier method.