The string parsing has been optimized to use ReadOnlySpan<char> instead of string in the ToFieldIdentifier method.

**BenchmarkDotNet Results - `propertyPath = "FirstName"`**

|               Method |     Mean |    Error |   StdDev | Allocated |
|--------------------- |---------:|---------:|---------:|----------:|
|    ToFieldIdentifier | 48.25 ns | 0.920 ns | 0.861 ns |         - |
|    ToFieldIdentifierNew | 45.81 ns | 0.411 ns | 0.384 ns |         - |



**BenchmarkDotNet Results - `propertyPath = "Address.Postcode"`**

|               Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|--------------------- |-----------:|---------:|---------:|-------:|----------:|
|    ToFieldIdentifier | 1,638.4 ns | 18.57 ns | 16.46 ns | 0.0248 |      80 B |
| ToFieldIdentifierNew |   196.6 ns |  1.36 ns |  1.27 ns | 0.0255 |      80 B |



**BenchmarkDotNet Results - `propertyPath = "Address.Keys[10].Name"`**

|               Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|--------------------- |-----------:|---------:|---------:|-------:|----------:|
|    ToFieldIdentifier | 4,862.3 ns | 70.15 ns | 58.58 ns | 0.1068 |     344 B |
| ToFieldIdentifierNew |   617.3 ns |  8.86 ns |  7.86 ns | 0.0706 |     224 B |